chore : 테스트 코드 분리

+++ src/main/java/step1/operater/Division.java
@@ -3,6 +3,6 @@ package step1.operater;
 public class Division implements Operater {
     @Override
     public int calculate(int operand1, int operand2) {
-        return (int) (operand1 / operand2);
+        return operand1 / operand2;
     }
 }
